requirements.txt
guessit>=0.9
pyxdg>=0.25
pony>=0.6
click>=3.3
